% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERVmap.R
\name{ERVmapParam}
\alias{ERVmapParam}
\title{Build an ERVmap parameter object}
\usage{
ERVmapParam(
  bfl,
  teFeatures,
  aggregateby = character(0),
  geneFeatures = NA,
  singleEnd = TRUE,
  ignoreStrand = TRUE,
  strandMode = 1L,
  fragments = !singleEnd,
  filterUniqReads = FALSE,
  maxMismatchRate = 0.02,
  suboptimalAlignmentTag = "auto",
  suboptimalAlignmentCutoff = 5
)
}
\arguments{
\item{bfl}{A \code{BamFile} or \code{BamFileList} object, or a character
string vector of BAM filenames. In order to apply all three filters present
in the ERVmap algorithm, reads must be aligned using Burrows-Wheeler Aligner.
If a different aligner was used, the 3rd filter (AS - XS >= 5) from the 
ERVmap algorithm is not applied.}

\item{teFeatures}{A \code{GRanges} or \code{GRangesList} object with the
TE annotated features to be quantified. Elements in this object should have
names, which will be used as a grouping factor for genomic ranges forming a
common locus, unless other metadata column names are specified in the
\code{aggregateby} parameter.}

\item{aggregateby}{Character vector with column names in the annotation
to be used to aggregate quantifications. By default, this is an empty vector,
which means that the names of the input \code{GRanges} or \code{GRangesList}
object given in the \code{annotations} parameter will be used to aggregate
quantifications.}

\item{geneFeatures}{A \code{GRanges} or \code{GRangesList} object with the
gene annotated features to be quantified. Only unique counts are used for
quantifying gene features given in this parameter.}

\item{singleEnd}{(Default TRUE) Logical value indicating if reads are single
(\code{TRUE}) or paired-end (\code{FALSE}).}

\item{ignoreStrand}{(Default TRUE) A logical which defines if the strand
should be taken into consideration when computing the overlap between reads
and TEs/ERVs in the annotations. When \code{ignore_strand = FALSE}, the
\code{\link[GenomicAlignments]{summarizeOverlaps}} function will only
consider those reads selected after filtering which overlap the TE or
ERV on the same strand. On the contrary, when \code{ignore_strand = TRUE},
the \code{\link[GenomicAlignments]{summarizeOverlaps}} function will count
any alignment which overlaps with the element in the annotations regardless
of the strand. For further details see
\code{\link[GenomicAlignments]{summarizeOverlaps}}.}

\item{strandMode}{(Default 1) Numeric vector which can take values 0, 1 or 2.
The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then \code{strandMode}
is ignored.}

\item{fragments}{(Default not \code{singleEnd}) A logical; applied to
paired-end data only. When \code{fragments=TRUE} (default), the read-counting
method in the original ERVmap algorithm will be applied, by which each mate
of a paired-end read is counted once, and therefore two mates mapping to the
same element result in adding up a count value of two. When
\code{fragments=FALSE}, if the two mates of a paired-end read map to the same
element, they are counted as a single hit and singletons, reads with unmapped
pairs and other fragments are not counted.}

\item{filterUniqReads}{(Default TRUE) Logical value indicating whether to apply
the alignment filters to unique reads (TRUE) or not (FALSE). These filters,
which are always applied to multi-mapping reads, can be optional for unique
reads, only if the NH tag is present in the BAM file. If 
\code{filterUniqReads = TRUE} (equivalent to the original approach proposed
by ERVmap authors), the unique reads not passing one or more filters 
from the ERVmap pipeline will be discarded to compute TEs expression.}

\item{maxMismatchRate}{(Default 0.02) Numeric value storing the maximum mismatch
rate employed by the ERVmap algorithm to discard aligned reads whose rate of
of sum of hard and soft clipping, or of the edit distance over the genome
reference, to the length of the read is above this threshold.}

\item{suboptimalAlignmentTag}{(Default "auto") Character string storing the
tag name in the BAM files that stores the suboptimal alignment score used in
the third filter of ERVmap; see Tokuyama et al. (2018). The default,
\code{suboptimalAlignmentTag="auto"}, assumes that either the BAM files were
generated by BWA and include a tag called \code{XS} that stores the suboptimal
alignment score, or if the \code{XS} tag is not available, then it will use
the available secondary alignments to implement an analogous approach to that
third ERVmap filter. When \code{suboptimalAlignmentTag="none"}, then it will
perform the latter approach also when the tag \code{XS} is avalable. When
this parameter is different from \code{"auto"} and \code{"none"}, then a tag
with the given name will be used to extract the suboptimal alignment score.
The absence of that tag will prompt an error.}

\item{suboptimalAlignmentCutoff}{(Default 5) Numeric value storing the cutoff
above which the difference between the alignment score and the suboptimal
alignment score is considered sufficiently large to retain the alignment.}
}
\value{
A \linkS4class{ERVmapParam} object.
}
\description{
Build an object of the class \code{ERVmapParam}
}
\details{
This is the constructor function for objects of the class
\code{ERVmapParam-class}. This type of object is the input to the
function \code{\link{qtex}()} for quantifying expression of transposable
elements using the ERVmap method
\href{https://doi.org/10.1073/pnas.1814589115}{Tokuyama et al. (2018)}. The
ERVmap algorithm processes reads following conservative filtering criteria
to provide reliable raw count data for each TE.
}
\examples{
bamfiles <- list.files(system.file("extdata", package="atena"),
                       pattern="*.bam", full.names=TRUE)
annot <- ERVmap_ann()
empar <- ERVmapParam(bamfiles, annot, singleEnd = TRUE)
empar

}
\references{
Tokuyama M et al. ERVmap analysis reveals genome-wide transcription of human
endogenous retroviruses. PNAS. 2018;115(50):12565-12572. DOI:
\url{https://doi.org/10.1073/pnas.1814589115}
}
