% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getLINEs}
\alias{getLINEs}
\title{Getter of LINE class TEs from parsed RepeatMasker annotations}
\usage{
getLINEs(parsed_ann, relLength = 0.9)
}
\arguments{
\item{parsed_ann}{A \link[GenomicRanges:GRangesList-class]{GRangesList} 
object obtained from parsing RepeatMasker annotations
with \code{OneCodeToFindThemAll} function.}

\item{relLength}{(Default 0.9) Numeric value that can take values between 0
to 1. Sets the minimum relative length required for
features. Elements with a lower relative length than
\code{relLength} will be filtered. The relative length
used is the one obtained by \code{OneCodeToFindThemAll}
(length of the reconstructed TE / length of the reference).}
}
\value{
A \link[GenomicRanges:GRangesList-class]{GRangesList} object with
        annotations from LINEs.
}
\description{
Getter of LINE class TEs from parsed RepeatMasker annotations
}
\details{
Retrieves LINE class TEs from RepeatMasker annotations after parsing using
the \code{OneCodeToFindThemAll} function. The \code{relLength} parameter
can be used to filter out elements with a lower relative length.
}
\examples{
rmsk_gr <- annotaTEs(genome = "dm6", parsefun = OneCodeToFindThemAll,
                     dictionary=NULL, fuzzy = FALSE, strict = FALSE)
rmsk_gr_line <- getLINEs(rmsk_gr, relLength = 0.95)
 
}
