% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSuboptimalAlignScore.R
\name{getSuboptimalAlignScore}
\alias{getSuboptimalAlignScore}
\title{Obtain the suboptimal alignment score tag}
\usage{
getSuboptimalAlignScore(r, strandMode = 1)
}
\arguments{
\item{r}{\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
object lacking the suboptimal alignment score tag and containing only
alignments corresponding to multi-mapping reads.}

\item{strandMode}{(Default 1) Numeric vector which can take values 0, 1 or 2. The strand
mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail.}
}
\value{
A \code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
  object with primary alignments of multi-mapping reads displaying the
  suboptimal alignment score tag ('XS') for both the first and last mate of
  the paired-end reads.
}
\description{
\code{getSuboptimalAlignScore} computes the suboptimal alignment score for
primary alignments by leveraging the information provided by
secondary alignments of multi-mapping reads in binary 'BAM' files. In
summary, the secondary alignments of each primary alignment are identified
and the alignment score of the best secondary alignment is reported as the
suboptimal alignment score. This function is called internally by
\code{\link{count_TE}} when the suboptimal alignment score tag is not
reported in the provided BAM files, since this information is required by
\code{\link{count_TE}} to correctly quantify the expression of transposable
elements.
}
\details{
This function aims at replacing the 'XS' tag reported by the
\href{http://bio-bwa.sourceforge.net/bwa.shtml}{Burrows-Wheeler Aligner}
(BWA), so that RNA-sequencing datasets aligned to the reference
genome using other software tools different to BWA can also take advantage
of this pipeline.
}
\examples{
\dontrun{path_bam_files_sorted <- "Documents/RNAseq/BAMfiles_sorted"
bam_files <- list.files(path_bam_files_sorted,
                        pattern = "\\\\.bam$",
                        full.names = TRUE)
bfl <- BamFileList(bam_files, asMates = TRUE)

sbflags <- scanBamFlag(isUnmappedQuery = FALSE,
                       isProperPair = TRUE,
                       isDuplicate = FALSE,
                       isNotPassingQualityControls = FALSE)

param <- ScanBamParam(flag = sbflags,
                      tag = c("NM", "AS", "NH", "XS"),
                      what = scanBamWhat(),
                      tagFilter = list(NH = c(2:10000)))

r <- readGAlignmentPairs(file, use.names=TRUE,
                         param=param, strandMode=2)

getSuboptimalAlignScore(r, strandMode=2)}

}
\seealso{
\code{\link[Rsamtools]{BamFile}},
\code{\link[Rsamtools]{scanBamFlag}},
\code{\link[Rsamtools]{ScanBamParam}},
\code{\link[GenomicAlignments]{readGAlignmentPairs}},
\code{\link[GenomicAlignments]{GAlignmentPairs-class}},
\code{\link{count_TE}}
}
