% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERVmap.R
\name{qtex,ERVmapParam-method}
\alias{qtex,ERVmapParam-method}
\alias{qtex}
\title{Quantify transposable element expression using ERVmap}
\usage{
\S4method{qtex}{ERVmapParam}(x)
}
\arguments{
\item{x}{An \code{ERVmapParam} object. It should be built
using the constructor function \code{\link{ERVmapParam}}.}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Using an \code{ERVmapParam} object as input, the
\code{qtex()} method quantifies the expression of transposable
elements (TEs) using the ERVmap algorithm. If you use this
method please cite Tokuyama et al. (2018).
}
\references{
Tokuyama M et al. ERVmap analysis reveals genome-wide transcription of human
endogenous retroviruses. PNAS. 2018;115(50):12565-12572. DOI:
\url{https://doi.org/10.1073/pnas.1814589115}
}
\seealso{
\code{\link{ERVmapParam}}
}
