% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{rmskidentity}
\alias{rmskidentity}
\title{Identity function for parsefun}
\usage{
rmskidentity(gr)
}
\arguments{
\item{gr}{A [`GRanges`][GenomicRanges::GRanges-class] object.}
}
\value{
A [`GRanges`][GenomicRanges::GRanges-class] object.
}
\description{
Identity function for parsefun
}
\details{
Identity function: returns the [`GRanges`][GenomicRanges::GRanges-class]
object without any modification.
}
\examples{
rmskid <- annotaTEs(genome="dm6", parsefun=rmskidentity)
rmskid
        
}
