% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_bam.R
\name{sort_bam}
\alias{sort_bam}
\title{Sort by Query-name all BAM files in directory}
\usage{
sort_bam(path_bam_files, path_bam_files_sorted)
}
\arguments{
\item{path_bam_files}{Character vector indicating the path in which the
binary ‘BAM’ files to be sorted by the query-name (read ID) are located.}

\item{path_bam_files_sorted}{Character vector indicating the location where
the resulting sorted BAM files will be created. This directory must be
different from \code{path_bam_files}, otherwise the unsorted 'BAM' files
will be overwritten.}
}
\value{
The function does not output any object. The resulting sorted files
  can be found in \code{path_bam_files_sorted}.
}
\description{
\code{sort_bam} allows to sort by Query-name (read ID) all binary 'BAM' files
located in the same directory. This function is based on the
\code{\link[Rsamtools]{sortBam}} function. This step is required
previous to the \code{\link{count_TE}} function when the suboptimal alignment
score tag is not specified in the 'BAM' files, in order to compute it.
}
\examples{
\dontrun{path_bam_files <- "Documents/RNAseq/BAMfiles"
path_bam_files_sorted <- "Documents/RNAseq/BAMfiles_sorted"
sort_bam(path_bam_files, path_bam_files_sorted)}

}
\seealso{
\code{\link{count_TE}},
\code{\link[Rsamtools]{sortBam}},
}
