% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/Telescope.R
\docType{class}
\name{TelescopeParam-class}
\alias{TelescopeParam-class}
\alias{show,TelescopeParam-method}
\title{Telescope parameter class}
\usage{
\S4method{show}{TelescopeParam}(object)
}
\description{
This is a class for storing parameters
provided to the Telescope algorithm.
}
\section{Slots}{

\describe{
\item{\code{basiliskEnv}}{A 'BasiliskEnvironment' object; see the
\code{BasiliskEnvironment-class} in the \code{basilisk} package.}

\item{\code{telescopeVersion}}{'character' string storing the Telescope version.}

\item{\code{telescopeOptions}}{'list' object storing the options for the
Telescope algorithm.}
}}

\references{
Bendall et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Comp. Biol. 2019;15(9):e1006453. DOI:
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
