% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Telescope.R
\name{TelescopeParam}
\alias{TelescopeParam}
\title{Build a Telescope parameter object}
\usage{
TelescopeParam(bfl, annotations, opts = list(quiet = TRUE))
}
\arguments{
\item{bfl}{A 'BamFile' or 'BamFileList' object, or a character string vector
of BAM filenames.}

\item{annotations}{A 'GRanges' object. Ranges in this object should have
names, which will be used as grouping factor for ranges forming a common locus.}

\item{opts}{Options to to pass to the telescope algorithm. Defaults
correspond to those from the telescope software, with the exception of
setting \code{quiet=TRUE} by using \code{--quiet} in the call to telescope.
The following Telescope options cannot be set: \code{--attribute},
\code{--outdir}, \code{--exp_tag}, \code{--tempdir}, \code{--ncpu},
\code{--skip_em}.}
}
\value{
A \linkS4class{TelescopeParam} object.
}
\description{
Build an object of the class \code{TelescopeParam-class}.
}
\details{
This is the constructor function for objects of the class
\code{TelescopeParam-class}. This type of object is the input to the
function \code{\link{qtex}()} for quantifying expression of transposable
elements, which will call the Telescope algorithm with this type of object.
}
\examples{
\dontshow{basilisk.utils::installConda()}
bamfiles <- list.files(system.file("extdata", package="atena"),
                       pattern="*.bam", full.names=TRUE)
annot <- ERVmap_ann()
tspar <- TelescopeParam(bamfiles, annot)
tspar

}
\references{
Bendall et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Comp. Biol. 2019;15(9):e1006453. DOI:
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
