% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotaTEs}
\alias{annotaTEs}
\title{Get RepeatMasker UCSC annotations}
\usage{
annotaTEs(genome = "hg38", parsefun = rmskidentity)
}
\arguments{
\item{genome}{The genome version of the desired RepeatMasker annotations
(e.g. "hg38").}

\item{parsefun}{A \code{function} to parse the annotations:
\itemize{
  \item Function \code{rmskidentity} returns RepeatMasker annotations as
        present in \link[AnnotationHub]{AnnotationHub}, without
        processing them.
  \item Function \code{rmskbasicparser} parses annotations by removing
        low complexity regions, simple repeats, satellites, rRNA, scRNA,
        snRNA, srpRNA and tRNA. Also removes TEs
        with a strand different than "+" or "-". Modifies "repFamily" and
        "repClass" columns when a "?" is present or when they are defined
        as "Unknown" or "Other". Finally, assigns a unique id to each TE
        instance by adding the suffix "_dup" plus a number at the end of
        the "repName".
  \item User-defined function. Input and output should be \code{GRanges}
        objects.
}}
}
\value{
A \link[GenomicRanges:GRanges-class]{GRanges} object with
        transposable element annotations.
}
\description{
The \code{annotaTEs()} function fetches RepeatMasker UCSC transposable
element (TE) annotations using 
\link[AnnotationHub]{AnnotationHub} and parses them.
}
\details{
Given a specific genome version, the \code{annotaTEs()} function fetches
RepeatMasker annotations from UCSC Genome Browser using the 
\link[AnnotationHub]{AnnotationHub} package. Since RepeatMasker not only
provides TE annotations but also low complexity DNA sequences and other
types of repeats, a specific \code{parsefun} can be set to parse these
annotations (e.g. \code{rmskbasicparser} or a user-defined function). If no
parsing is required, \code{parsefun} can be set to \code{rmskidentity}.
}
\examples{
rmsk_gr <- annotaTEs(genome = "hg38", parsefun = rmskidentity)


}
\seealso{
\code{\link[AnnotationHub]{AnnotationHub}}
}
