% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AtenaParam-class}
\alias{AtenaParam-class}
\alias{path,AtenaParam-method}
\title{Atena parameter class}
\usage{
\S4method{path}{AtenaParam}(object)
}
\description{
This is a virtual class from which other classes are derived
for storing parameters provided to quantification methods of
transposable elements from RNA-seq data.
}
\section{Slots}{

\describe{
\item{\code{bfl}}{A 'BamFileList' object.}

\item{\code{annotations}}{A 'GRanges' object.}
}}

\seealso{
\code{\link{ERVmapParam-class}}
\code{\link{TelescopeParam-class}}
}
