% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AtenaParam-class}
\alias{AtenaParam-class}
\title{Atena parameter class}
\description{
This is a top-level class for storing parameters
provided to several functions in the atena package.
}
\section{Slots}{

\describe{
\item{\code{bfl}}{A 'BamFileList' object.}

\item{\code{annotations}}{A 'GRanges' object.}
}}

