% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AtenaParam-class}
\alias{AtenaParam-class}
\title{Atena parameter class}
\description{
This is a top-level class for storing parameters
provided to several functions in the atena package.
}
\section{Slots}{

\describe{
\item{\code{bfl}}{A 'BamFileList' object.}

\item{\code{annotations}}{A 'GRanges' object.}
}}

\references{
Bendall et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Comp. Biol. 2019;15(9):e1006453. DOI:
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
