% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AtenaParam-class}
\alias{AtenaParam-class}
\alias{path,AtenaParam-method}
\title{Atena parameter class}
\usage{
\S4method{path}{AtenaParam}(object)
}
\arguments{
\item{object}{A \linkS4class{AtenaParam} object.}
}
\description{
This is a virtual class from which other classes are derived
for storing parameters provided to quantification methods of
transposable elements from RNA-seq data.
}
\section{Slots}{

\describe{
\item{\code{bfl}}{A \linkS4class{BamFileList} object.}

\item{\code{annotations}}{A \linkS4class{GRanges} object.}

\item{\code{aggregateby}}{Character vector with column names in the annotation
to be used to aggregate quantifications.}
}}

\seealso{
\code{\link{ERVmapParam-class}}
\code{\link{TelescopeParam-class}}
\code{\link{TEtranscriptsParam-class}}
}
