% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AtenaParam-class}
\alias{AtenaParam-class}
\alias{path,AtenaParam-method}
\alias{features,AtenaParam-method}
\alias{features}
\title{Atena parameter class}
\usage{
\S4method{path}{AtenaParam}(object)

\S4method{features}{AtenaParam}(object)
}
\arguments{
\item{object}{A \linkS4class{AtenaParam} object.}
}
\value{
\code{path()}: Filesystem paths to the BAM files in the input
parameter object.

code{features()}: The \code{GenomicRanges} or
\code{GenomicRangesList} object with the features in the input parameter
object.
}
\description{
This is a virtual class from which other classes are derived
for storing parameters provided to quantification methods of
transposable elements from RNA-seq data.
}
\section{Slots}{

\describe{
\item{\code{bfl}}{A \linkS4class{BamFileList} object.}

\item{\code{features}}{A \linkS4class{GRanges} object.}

\item{\code{aggregateby}}{Character vector with column names in the annotation
to be used to aggregate quantifications.}
}}

\examples{
bamfiles <- list.files(system.file("extdata", package="atena"),
                       pattern="*.bam", full.names=TRUE)
annot <- ERVmap_ann()
ttpar <- TEtranscriptsParam(bfl = bamfiles, teFeatures = annot, singleEnd = TRUE)
path(ttpar)

}
\seealso{
\code{\link{ERVmapParam-class}}
\code{\link{TelescopeParam-class}}
\code{\link{TEtranscriptsParam-class}}
}
