% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ERVmap.R
\docType{class}
\name{ERVmapParam-class}
\alias{ERVmapParam-class}
\alias{show,ERVmapParam-method}
\title{ERVmap parameter class}
\usage{
\S4method{show}{ERVmapParam}(object)
}
\arguments{
\item{object}{A \linkS4class{ERVmapParam} object.}
}
\description{
This is a class for storing parameters provided to the ERVmap algorithm.
It is a subclass of the 'AtenaParam-class'.
}
\section{Slots}{

\describe{
\item{\code{singleEnd}}{(Default FALSE) Logical value indicating if reads are single
(\code{TRUE}) or paired-end (\code{FALSE}).}

\item{\code{strandMode}}{(Default 1) Numeric vector which can take values 0, 1 or 2.
The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then use either
\code{strandMode = NULL} or do not specify the \code{strandMode} parameter.}

\item{\code{ignoreStrand}}{(Default TRUE) A logical which defines if the strand
should be taken into consideration when computing the overlap between reads
and TEs/ERVs in the annotations. When \code{ignore_strand = FALSE}, the
\code{\link[GenomicAlignments]{summarizeOverlaps}} function will only
consider those reads selected after filtering which overlap the TE or
ERV on the same strand. On the contrary, when \code{ignore_strand = TRUE},
the \code{\link[GenomicAlignments]{summarizeOverlaps}} function will count
any alignment which overlaps with the element in the annotations regardless
of the strand. For further details see
\code{\link[GenomicAlignments]{summarizeOverlaps}}.}

\item{\code{filterUniqReads}}{(Default TRUE) Logical value indicating whether to apply
the alignment filters to unique reads (TRUE) or not (FALSE). These filters,
which are always applied to multi-mapping reads, are optional for unique
reads. If TRUE, the unique reads not passing one or more filters from the
ERVmap pipeline, except for the "AS - XS >= 5" filter, will be discarded to
compute TEs expression.}

\item{\code{fragments}}{(Default TRUE) A logical; applied to paired-end data only.
When \code{fragments=TRUE} (default), the read-counting method in the
original ERVmap algorithm will be applied, by which each mate of a paired-end
read is counted once, and therefore two mates mapping to the same element
result in adding up a count value of two. When \code{fragments=FALSE}, if the
two mates of a paired-end read map to the same element, they are counted as
a single hit and singletons, reads with unmapped pairs and other fragments
are not counted.}
}}

\references{
Tokuyama M et al. ERVmap analysis reveals genome-wide transcription of human
endogenous retroviruses. PNAS. 2018;115(50):12565-12572. DOI:
\url{https://doi.org/10.1073/pnas.1814589115}
}
