% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERVmap.R, R/Telescope.R
\name{qtex,ERVmapParam-method}
\alias{qtex,ERVmapParam-method}
\alias{qtex}
\alias{qtex,TelescopeParam-method}
\title{Quantify transposable element expression}
\usage{
\S4method{qtex}{ERVmapParam}(x)

\S4method{qtex}{TelescopeParam}(x, phenodata = NULL, BPPARAM = SerialParam(progress = TRUE))
}
\arguments{
\item{x}{A \code{TelescopeParam} object containing the configuration
parameters to call the Telescope algorithm.}

\item{phenodata}{A \code{data.frame} or \code{DataFrame} object storing
phenotypic data to include in the resulting
\code{SummarizedExperiment} object. If \code{phenodata} is set,
its row names will become the column names of the resulting
\linkS4class{SummarizedExperiment} object.}

\item{BPPARAM}{An object of a \linkS4class{BiocParallelParam} subclass
to configure the parallel execution of the code. By default,
a \linkS4class{SerialParam} object is used, which does not use
any parallelization, with the flag \code{progress=TRUE} to show
progress through the calculations.}
}
\value{
A \code{SummarizedExperiment} object.

A \linkS4class{SummarizedExperiment} object.
}
\description{
Giving an \code{AtenaParam} object as input, the
\code{qtex()} method quantifies the expression of transposable
elements (TEs). The particular algorithm to perform the
quantification will be selected depending on the specific
sub-class of input \code{AtenaParam} object, see argument
\code{x} below.
}
\references{
Tokuyama M et al. ERVmap analysis reveals genome-wide transcription of human
endogenous retroviruses. PNAS, 115(50):12565-12572, 2018.
\url{https://doi.org/10.1073/pnas.1814589115}

Bendall ML et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Computational Biology, 15:e1006453, 2019.
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
\seealso{
\code{\link{ERVmapParam}}
}
