% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERVmap.R, R/Telescope.R
\name{qtex,ERVmapParam-method}
\alias{qtex,ERVmapParam-method}
\alias{qtex}
\alias{qtex,TelescopeParam-method}
\title{Quantify transposable element expression}
\usage{
\S4method{qtex}{ERVmapParam}(x)

\S4method{qtex}{TelescopeParam}(x, BPPARAM = SerialParam())
}
\arguments{
\item{x}{An \code{AtenaParam} object of one of the following
subclasses:
\itemize{
  \item A \code{ERVmapParam} object built using the constructor
        function \code{\link{ERVmapParam}()}. This object will
        trigger \code{qtex()} to use the algorithm by
        Tokuyama et al. (2018).
  \item A \code{TelescopeParam} object built using the constructor
        function \code{\link{TelescopeParam}()}. This object will
        trigger \code{qtex()} to use the algorithm by
        Bendall et al. (2019).
}}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Giving an \code{AtenaParam} object as input, the
\code{qtex()} method quantifies the expression of transposable
elements (TEs). The particular algorithm to perform the
quantification will be selected depending on the specific
sub-class of input \code{AtenaParam} object, see argument
\code{x} below.
}
\references{
Tokuyama M et al. ERVmap analysis reveals genome-wide transcription of human
endogenous retroviruses. PNAS, 115(50):12565-12572, 2018.
\url{https://doi.org/10.1073/pnas.1814589115}

Bendall ML et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Computational Biology, 15:e1006453, 2019.
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
\seealso{
\code{\link{ERVmapParam}}
}
