% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/Telescope.R
\docType{class}
\name{TelescopeParam-class}
\alias{TelescopeParam-class}
\alias{show,TelescopeParam-method}
\title{Telescope parameter class}
\usage{
\S4method{show}{TelescopeParam}(object)
}
\arguments{
\item{object}{A \linkS4class{TelescopeParam} object.}
}
\description{
This is a class for storing parameters
provided to the Telescope algorithm.
}
\section{Slots}{

\describe{
\item{\code{singleEnd}}{(Default TRUE) Logical value indicating if reads are single
(\code{TRUE}) or paired-end (\code{FALSE}).}

\item{\code{strandMode}}{(Default 1) Numeric vector which can take values 0, 1 or 2.
The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then \code{strandMode}
is ignored.}

\item{\code{ignoreStrand}}{(Default FALSE) A logical which defines if the strand
should be taken into consideration when computing the overlap between reads
and annotated features. When \code{ignoreStrand = FALSE}, an aligned read
is considered to be overlapping an annotated feature as long as they
have a non-empty intersecting genomic range on the same strand, while when
\code{ignoreStrand = TRUE} the strand is not considered.}

\item{\code{fragments}}{(Default FALSE) A logical; applied to paired-end data only.
When \code{fragments=FALSE} (default), the read-counting method only counts 
‘mated pairs’ from opposite strands, while when \code{fragments=TRUE},
same-strand pairs, singletons, reads with unmapped pairs and other fragments 
are also counted. For further details see
\code{\link[GenomicAlignments]{summarizeOverlaps}()}.}

\item{\code{pi_prior}}{(Default 0) A positive integer scalar indicating the prior 
on pi. This is equivalent to adding n unique reads.}

\item{\code{theta_prior}}{(Default 0) A positive integer scalar storing the prior 
on Q. Equivalent to adding n non-unique reads.}

\item{\code{em_epsilon}}{(Default 1e-7) A numeric scalar indicating the EM 
Algorithm Epsilon cutoff.}

\item{\code{maxIter}}{A positive integer scalar storing the maximum number of
iterations of the EM SQUAREM algorithm (Du and Varadhan, 2020). Default
is 100 and this value is passed to the \code{maxiter} parameter of the
\code{\link[SQUAREM]{squarem}()} function.}
}}

\references{
Bendall et al. Telescope: characterization of the retrotranscriptome by
accurate estimation of transposable element expression.
PLOS Comp. Biol. 2019;15(9):e1006453. DOI:
\url{https://doi.org/10.1371/journal.pcbi.1006453}
}
