% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TelescopeParam-class}
\alias{TelescopeParam-class}
\title{Telescope parameter class}
\description{
This is a class for storing parameters
provided to the Telescope algorithm.
}
\section{Slots}{

\describe{
\item{\code{basiliskEnv}}{A 'BasiliskEnvironment' object; see the
\code{BasiliskEnvironment-class} in the \code{basilisk} package.}

\item{\code{telescopeOptions}}{'list' object storing the options for the
Telescope algorithm.}
}}

