% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/TEtranscripts.R
\docType{class}
\name{TEtranscriptsParam-class}
\alias{TEtranscriptsParam-class}
\alias{show,TEtranscriptsParam-method}
\title{TEtranscripts parameter class}
\usage{
\S4method{show}{TEtranscriptsParam}(object)
}
\arguments{
\item{object}{A \linkS4class{TEtranscriptsParam} object.}
}
\description{
This is a class for storing parameters provided to the TEtranscripts algorithm.
It is a subclass of the 'AtenaParam-class'.
}
\section{Slots}{

\describe{
\item{\code{singleEnd}}{(Default FALSE) Logical value indicating if reads are single
(\code{TRUE}) or paired-end (\code{FALSE}).}

\item{\code{ignoreStrand}}{(Default FALSE) A logical which defines if the strand
should be taken into consideration when computing the overlap between reads
and annotated features. When \code{ignoreStrand = FALSE}, an aligned read
will be considered to be overlapping an annotated feature as long as they
have a non-empty intersecting genomic ranges on the same strand, while when
\code{ignoreStrand = TRUE} the strand will not be considered.}

\item{\code{strandMode}}{(Default 1) Numeric vector which can take values 0, 1 or 2.
The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then use either
\code{strandMode = NULL} or do not specify the \code{strandMode} parameter.}

\item{\code{fragments}}{(Default TRUE) A logical; applied to paired-end data only.
When \code{fragments=TRUE} (default), the read-counting method will also
count reads without mates, while when \code{fragments=FALSE} those reads
will not be counted. For further details see
\code{\link[GenomicAlignments]{summarizeOverlaps}()}.}

\item{\code{tolerance}}{A positive numeric scalar storing the minimum tolerance
above which the SQUAREM algorithm (Du and Varadhan, 2020) keeps iterating.
Default is \code{1e-4} and this value is passed to the \code{tol} parameter
of the \code{\link[SQUAREM]{squarem}()} function.}

\item{\code{maxIter}}{A positive integer scalar storing the maximum number of
iterations of the SQUAREM algorithm (Du and Varadhan, 2020). Default
is 100 and this value is passed to the \code{maxiter} parameter of the
\code{\link[SQUAREM]{squarem}()} function.}
}}

\references{
Jin Y et al. TEtranscripts: a package for including transposable elements
in differential expression analysis of RNA-seq datasets.
Bioinformatics. 2015;31(22):3593-3599. DOI:
\url{https://doi.org/10.1093/bioinformatics/btv422}
}
